Added .travis.yml (#9)
This PR enables us to check that

the code base sent as a PR can be compiled correctly
$ gradlew check works properly (and this means tests are green)
Additionally, I changed the settings of repositories (drive/drivecommand). This makes following changes:

when the test fails, merge button will be disabled.
when owners (PILE members) push some code to develop or master, the patch will be merged into the branch after testing. If the test fails, merge process will be canceled. (I've not checked it yet, though...)

Issue #8